Takes the exact declared variable as the parameter of the functions

+++ contracts/test/LON/Burn.t.sol
@@ -7,7 +7,7 @@ contract TestLONBurn is TestLON {
     using BalanceSnapshot for BalanceSnapshot.Snapshot;
 
     function testCannotBurnMoreThanOwned() public {
-        lon.mint(user, 1e18);
+        lon.mint(user, uint256(1e18));
         uint256 excessAmount = lon.balanceOf(user) + 1;
         vm.expectRevert("ERC20: burn amount exceeds balance");
         vm.prank(user);
@@ -15,7 +15,7 @@ contract TestLONBurn is TestLON {
     }
 
     function testBurn() public {
-        lon.mint(user, 1e18);
+        lon.mint(user, uint256(1e18));
         BalanceSnapshot.Snapshot memory userLon = BalanceSnapshot.take(user, address(lon));
         vm.prank(user);
         lon.burn(1e18);
