+++ contracts/test/LON/EmergencyWithdraw.t.sol
@@ -8,8 +8,8 @@ contract TestLONEmergencyWithdraw is TestLON {
     using BalanceSnapshot for BalanceSnapshot.Snapshot;
 
     function testEmergencyWithdraw() public {
-        MockERC20 dai = new MockERC20("DAI", "DAI", 18);
-        dai.mint(address(lon), 1e18);
+        MockERC20 dai = new MockERC20("DAI", "DAI", uint8(18));
+        dai.mint(address(lon), uint256(1e18));
         BalanceSnapshot.Snapshot memory lonDai = BalanceSnapshot.take(address(lon), address(dai));
         BalanceSnapshot.Snapshot memory emergencyRecipientDai = BalanceSnapshot.take(emergencyRecipient, address(dai));
         vm.prank(user);
